requirements.txt
flask
flask-login
flask-sqlalchemy
psycopg2-binary

pandas
numpy
scikit-learn
matplotlib
seaborn

torch
transformers
tqdm

tensorflow
keras

beautifulsoup4
requests
selenium

jupyter
notebook
